+++ rest_api/src/main/java/spring/model/Role.java
@@ -7,12 +7,14 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Entity
 @Table(name = "roles")
 @Data
+@Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class Role {
