+++ rest_api/src/main/java/spring/model/auth/SignInRequestDTO.java
@@ -1,8 +1,10 @@
-package spring.model.dto;
+package spring.model.auth;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
+@AllArgsConstructor
 public class SignInRequestDTO {
     private String username;
     private String password;
