+++ rest_api/src/main/java/spring/model/auth/SingInResponseDTO.java
@@ -1,0 +1,12 @@
+package spring.model.auth;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class SingInResponseDTO {
+    String accessToken;
+}
